Add simple application logic and GET routes
The application logic is simple. We have a mock database named
"carrinho", which represents a shopping cart with items. The routes will
list the shopping cart contents, add more items to the shopping cart, up
date the entire cart or only a item of it and delete an item from the ca
rt.

Added routes to parse GET requests, as follows:
- GET route to list all items in a shopping cart
- GET route to list a single item of a given ID